Split bin_user and bin_prefix implementations
Module-level logic is bad.
